+++ app/Http/Controllers/Dashboard/EventController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Dashboard\EventRequest;
 use App\Models\Event;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
@@ -12,6 +13,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Validator;
 
 class EventController extends Controller
 {
@@ -39,18 +41,17 @@ class EventController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param Request $request
+     * @param EventRequest $request
      * @return RedirectResponse
      */
-    public function store(Request $request): RedirectResponse
+    public function store(EventRequest $request): RedirectResponse
     {
-        if ($request->hasfile('image')) {
-            $name = $request->image->getClientOriginalName();
-            $request->image->move(public_path() . '/events/' . $request->title . '/', $name);
-            $image = $name;
-        }
-
         try {
+            if ($request->hasfile('image')) {
+                $name = $request->image->getClientOriginalName();
+                $request->image->move(public_path() . '/events/' . $request->title . '/', $name);
+                $image = $name;
+            }
             $event = new event();
             $event->title = $request->title;
             $event->date = $request->date;
@@ -101,25 +102,48 @@ class EventController extends Controller
      */
     public function update(Request $request, Event $event): RedirectResponse
     {
-
-        $event->update($request->all());
-        if ($request->image != '') {
-            $path = public_path() . '/events/';
-
-            //code for remove old file
-            if ($event->image != '' && $event->image != null) {
-                $file_old = $path . $event->title . '/' . $event->image;
-//                dd($file_old);
-                Storage::disk('public')->delete($file_old);
+        try {
+            $validate = Validator::make($request->all(), [
+                'title'         => 'required|min:5',
+                'date'          => 'required|date|after:tomorrow',
+                'time'          => 'required',
+                'desc'          => 'required',
+                'image'         => 'mimes:jpeg,jpg,png',
+            ],[
+                'image.mimes'       =>'jpeg or jpg or png'
+            ]);
+            if ($validate->fails()){
+                return back()->withErrors($validate->errors())->withInput();
             }
-            //upload new file
-            $name = $request->image->getClientOriginalName();
-            $request->image->move(public_path() . '/events/' . $request->title . '/', $name);
-            $event->update(['image' => $name]);
+
+            if ($event->update($request->all())){
+                if ($request->image != '') {
+                    $path = public_path() . '/events/';
+
+                    //code for remove old file
+                    if ($event->image != '' && $event->image != null) {
+                        $file_old = $path . $event->title . '/' . $event->image;
+//                dd($file_old);
+                        Storage::disk('public')->delete($file_old);
+                    }
+                    //upload new file
+                    $name = $request->image->getClientOriginalName();
+                    $request->image->move(public_path() . '/events/' . $request->title . '/', $name);
+                    $event->update(['image' => $name]);
+                }
+
+                return redirect()->route('dashboard.events.index')
+                    ->with(['success' => __('global.success_update')]);
+            }
+            else{
+                return redirect()->route('dashboard.events.index')
+                    ->with(['error' => __('global.error_update')]);
+            }
         }
-
-            return redirect()->route('dashboard.events.index');
-
+        catch (\Exception $e){
+            return redirect()->route('dashboard.events.index')
+                ->with(['error' => __('global.data_error')]);
+        }
     }
 
     /**
@@ -130,8 +154,13 @@ class EventController extends Controller
      */
     public function destroy(Event $event): RedirectResponse
     {
-        $event->forceDelete();
-        return redirect()->route('dashboard.events.index');
+        if ($event->forceDelete()){
+            return redirect()->route('dashboard.events.index')
+                ->with(['success' => __('global.success_force_delete')]);
+        }
+        else
+            return redirect()->route('dashboard.events.index')
+                ->with(['error' => __('global.error_force_delete')]);
     }
 
     public function changeStatus(Event $event): RedirectResponse
